- Refactor callback sender - Introduce callback namespace
