+++ src/server/src/Stub/ActionController.php
@@ -7,8 +7,8 @@ namespace App\Stub;
 use App\Service\WebControllerService;
 use App\Stub\Collection\ArrayCollection;
 use App\Stub\Service\ActionFactory;
-use App\Stub\Service\CallbackProcessor;
-use App\Stub\Service\CallbackResolver;
+use App\Stub\Service\Callback\CallbackProcessor;
+use App\Stub\Service\Callback\CallbackResolver;
 use App\Stub\Session\State;
 use App\Stub\Session\StateManager;
 use LogicException;
